Add product and category management routes and handlers

+++ internal/utils/response.go
@@ -19,10 +19,10 @@ type PaginatedResponse struct {
 }
 
 type PaginationMeta struct {
-	Page       int `json:"page"`
-	Limit      int `json:"limit"`
-	Total      int `json:"total"`
-	TotalPages int `json:"total_pages"`
+	Page       int   `json:"page"`
+	Limit      int   `json:"limit"`
+	Total      int64 `json:"total"`
+	TotalPages int   `json:"total_pages"`
 }
 
 func SuccessResponse(c *gin.Context, message string, data interface{}) {
